Enable inside-out and outside-in bias sweeps
Add an option "Acquisition order" that determines the order in which the
points are acquired.  This option is only available when "Acquisition
mode" is set to "between points" (with or without lock-in).

"Acquisition order" may be:
- increasing: sweep bias from negative to positive
- decreasing: sweep bias from positive to negative
- inside-out: sweep bias from 0 to finite to measure switching currents
- outside-in: sweep bias from finite to 0 to measure retrapping currents

If inside-out or outside-in is selected, the number of points must be
even so that the positive and negative sides of the VI curve have an
equal number of points.  This condition is enforced by the driver when
either "Acquisition order" or "DMM: number of points" is set, as well as
when the driver is started.