Fix lock-in x-values of VI trace
The lock-in trace is acquired simultaneously with the DMM "VI curve" and
stored for later when performGetValue is called on "dR vs I curve".
This is why "VI curve" must always be acquired before "dR vs I curve".

We now store the correct x-values for the lock-in "dR vs I curve" which
were previously assumed to be always increasing.