Set integration time based on NPLC
Number of power-line cycles is agnostic to the power-line frequency
(although the high_lim assumes a 60Hz power-line frequency).  The
previous options were based on the (incorrect, in our case) assumption
of a 50Hz power-line frequency, and did not provide the user full
control to set the NPLC anywhere between 0.0005 and 15.