Provide direct control of DMM's NPLC and filter
The driver was previously designed to set the NPLC and filter values
based on the lock-in settling time, but the logic never worked for
arbitrary settling times and time constants.  Moreover the NPLC would be
set to 0.1, 1, or 10 despite the fact that the Keithley6500 is optimized
for 1-5 PLC and can take values between 0.0005 and 15.

The DMM's NPLC and filter settings can now be controlled directly, and
the lock-in settling time is provided as a read-only setting.  This
makes more sense because the settling time has always been determined by
the time required for the DMM to acquire a data point.

Support for the Keithley 2000 has been removed for now since it was
never used anyway.